Modification du controller sauce pour ajout verification ID sur modify et delete

+++ backend/controllers/sauce.js
@@ -22,20 +22,37 @@ exports.modifySauce = (req, res, next) => {
         ...JSON.parse(req.body.sauce),
         imageUrl : `${req.protocol}://${req.get("host")}/images/${req.file.filename}`
     } : { ...req.body };
-    Sauce.updateOne({ _id : req.params.id }, { ...sauceObject, _id : req.params.id })
-        .then(() => res.status(200).json({ message : "Sauce modifié" }))
-        .catch(error => res.status(400).json({error}))
+
+    Sauce.findOne({ _id : req.params.id })
+        .then(sauce => {
+            console.log(sauce.userId);
+            if (req.body.userId === sauce.userId) {
+                Sauce.updateOne({ _id : req.params.id }, { ...sauceObject, _id : req.params.id })
+                    .then(() => res.status(200).json({ message : "Sauce modifiée" }))
+                    .catch(error => res.status(400).json({ error }))
+            }
+            else {
+                res.status(404).json({ message : "Sauce non modifiée"})
+            }
+        })
+        .catch(error => res.status(500).json({ error }))
 };
 
 exports.deleteSauce = (req, res, next) => {
     Sauce.findOne({ _id : req.params.id })
     .then(sauce => {
-        const filename = sauce.imageUrl.split("/images/")[1];
-        fs.unlink(`images/${filename}`, () => {
-            Sauce.deleteOne({ _id : req.params.id })
-                .then(() => res.status(200).json({ message : "Sauce supprimé" }))
-                .catch(error => res.status(400).json({ error }))
-        })
+        console.log(sauce.userId);
+        if (req.body.userId === sauce.userId) {
+            const filename = sauce.imageUrl.split("/images/")[1];
+            fs.unlink(`images/${filename}`, () => {
+                Sauce.deleteOne({ _id : req.params.id })
+                    .then(() => res.status(200).json({ message : "Sauce supprimée" }))
+                    .catch(error => res.status(400).json({ error }))
+            })
+        }   
+        else {
+            res.status(404).json({ message : "Sauce non supprimée" })
+        }
     })
     .catch(error => res.status(500).json({ error }))
 };
